+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 - check_motor_disabled decorator does not work with positional arguments
 - Adapt get_encoded_image_from_dictionary for COM-KIT
 
+### Changed
+- Use ingenialink to get the drive's encoded image from the dictionary.
+
 ## [0.6.3] - 2023-10-11
 ### Fixed
 - Remove disturbance data before including new data.